apply latest rest-client code.
